+++ lib/restart-trigger.ts
@@ -1,2 +1,2 @@
-// restart triggered at 2026-03-05T08:48:12.918Z
-export const RESTART_TS = 1772700492918;
+// restart triggered at 2026-03-08T10:24:58.217Z
+export const RESTART_TS = 1772965498217;